Create esmvaltool._config module and close #225
